iota_json_rpc

Module coin_api

Structs

CoinReadApi

CoinReadInternalImpl

Traits

CoinReadInternal

CoinReadInternal trait to capture logic of interactions with AuthorityState and metrics This allows us to also mock internal implementation for testing

Functions

parse_to_struct_tag

parse_to_type_tag