iota_json_rpc::governance_api

Function mean_apy_from_exchange_rates

pub fn mean_apy_from_exchange_rates<'er>(
    exchange_rates: impl DoubleEndedIterator<Item = &'er PoolTokenExchangeRate> + Clone,
) -> f64

Calculate the APY using a 7-epoch moving average.

Returns the Mean by default, but falls back to the Median if outliers are detected. Outliers are defined as any APY > MAX_VALID_APY (100%) or if the trailing 8th epoch exchange rate is missing. This fallback protects against skewed results caused by large staking events or the spikes seen after missing exchange rates.