iota_json_rpc::governance_api

Function median_apy_from_exchange_rates

pub fn median_apy_from_exchange_rates<'er>(
    exchange_rates: impl DoubleEndedIterator<Item = &'er PoolTokenExchangeRate> + Clone,
) -> f64

Calculate the APY for a validator based on the exchange rates of the staking pool.

The calculation uses the median value of the sample, to filter out outliers introduced by large staking/unstaking events.